Added multiple catch series inputs
Allowed the imputation of multiple catch series. Multiple catch time series are now input as a single data.frame with different columns for different catch series. The struck and loss rates are then multiplied by the respective order of inclusion. The function also saves a time series of catch trajectories with added uncertainty.